[D3D11] Fix typo when skipping padding rows
The existing code has a closing parenthesis in the wrong place so the
final multiplication is outside of the macro.

This was noticed as part of fixing issues found by enabling
bugprone-implicit-widening-of-multiplication-result in clang-tidy.

It has been broken out into its own CL from the more mechanical/cleanup
fixes, since I suspected this would have functional impact.

From the review, because of how the underlying macro is written the
operation wouldn't be a no-op.

Fixes: 534377130
Reviewed-on: https://dawn-review.googlesource.com/c/dawn/+/324055
Commit-Queue: Ryan Harrison <rharrison@chromium.org>
Auto-Submit: Ryan Harrison <rharrison@chromium.org>

+++ src/dawn/native/d3d11/TextureD3D11.cpp
@@ -843,7 +843,7 @@ MaybeError Texture::WriteDepthStencilInternal(const ScopedCommandRecordingContex
         commandContext->Unmap(stagingTexture->GetD3D11Resource(), layer);
         DAWN_ASSERT(size.height <= rowsPerImage);
         // Skip the padding rows.
-        DAWN_UNSAFE_TODO(pSrcData += (rowsPerImage - size.height)) * bytesPerRow;
+        DAWN_UNSAFE_TODO(pSrcData += static_cast<size_t>(rowsPerImage - size.height) * bytesPerRow);
     }
 
     // Copy to the dest texture from the staging texture.
